New version:
Big change: added an option to do likelihood-analysis on the mock spectrum. This should give even tighter constraints given the additional information. There is (for now) one flag that controls it, and it's 'like_on_flux'. There are some implicit assumptions that will likely become new flags in the future.

Smaller changes:
likelihood is now given in log for numerical reasons that happend when a lot of galaxies are used.
Redshift of mocks is calculated more consistently though some more work will be needed.