Saving stuff beforehand so the container is not super-filled.
More details about the commit here: There was a problem with memory usage which ended on of the jobs. The main culprit is the memory usage of the container (I believe) so instead of filling the container with everything, I decided to save what needs to be save in the cache beforehand and saving only the final products afterwards.